+++ CHANGELOG.md
@@ -33,6 +33,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Operations for data cleanup
 - Examples of using data cleanup operations
   https://intel.github.io/dffml/examples/data_cleanup/index.html
+- Dev CMD to remove unused imports, `$ dffml service dev lint imports`
 ### Changed
 - Calls to hashlib now go through helper functions
 - Build docs using `dffml service dev docs`